Encapsulate robot logic within the Robot class
