+++ gherkin/model/__init__.py
@@ -1,5 +1,5 @@
 from gherkin.model.common import Angle, DIRECTION, Goal, SPEED, Rotation
-from gherkin.model.robot import Robot, RobotLimits
+from gherkin.model.robot import Robot
 from gherkin.model.world import World
 
-__all__ = ["Angle", "DIRECTION", "Goal", "Robot", "RobotLimits", "Rotation", "SPEED", "World"]
+__all__ = ["Angle", "DIRECTION", "Goal", "Robot", "Rotation", "SPEED", "World"]
