+++ gherkin/util/__init__.py
@@ -1,5 +1,4 @@
-from gherkin.util.controller import Controller
-from gherkin.util.runner import Runner
 from gherkin.util.utilities import generate_random_goal
+from gherkin.util.visualizer import Visualizer
 
-__all__ = ["Controller", "Runner", "generate_random_goal"]
+__all__ = ["generate_random_goal", "Visualizer"]
