--- gherkin/visuals/__init__.py
@@ -1,3 +1,0 @@
-from gherkin.visuals.visualizer import Visualizer
-
-__all__ = ["Visualizer"]
